# Prevent the parser from triggering a stack overflow