This makes the parser work with buffers containing thousands of frames.

Fixes #990.

cc: @Nibbler999